+++ src-tauri/src/main.rs
@@ -2,5 +2,5 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-    ai_tool_manager_lib::run()
+    ai_toolkit_lib::run()
 }